Add install docs
This package doesn't seem to follow the somewhat standard practice of using /cmd/${folder}/ for installing binary packages. This is fine it just means that to quickly pull the binary locally with go you need to use github.com/google/go-jsonnet/jsonnet as the path.

This will remove the issue of having to do anything else such as full fatih/color as well.