fix: Make lstripChars and rStripChars work with multibyte characters
We're currently working with the length of the string itself, which is
the number of bytes. That's wrong in these functions which operate at
the level of characters. When applied to a string with a multibyte
character they can return the wrong result. For example, evaluating
`std.rstripChars('• foo\n', '\n')` results in `"• f"`.

What we need to do instead is get the number of runes, and remove one
rune at a time.

Fixes `std.lstripChars` and `std.rstripChars` when the source string
contains multibyte (e.g. Unicode) characters.